posts and blogs with raw sql(4)

+++ src/domains/blogs/use-cases/delete-post-with-checking-use-case.ts
@@ -25,9 +25,10 @@ export class DeletePostWithCheckingUseCase
     if (!isUUID(command.blogId)) return HttpStatus.NOT_FOUND;
     if (!isUUID(command.postId)) return HttpStatus.NOT_FOUND;
 
-    // const blog = await this.blogsSqlRepository.fetchAllBlogDataById(
-    //   command.blogId,
-    // );
+    const blog = await this.blogsSqlRepository.fetchAllBlogDataById(
+      command.blogId,
+    );
+    if (!blog) return HttpStatus.NOT_FOUND;
     // if (blog && blog.userId !== command.userId) return HttpStatus.FORBIDDEN;
 
     const result = await this.postsSqlRepository.deletePost(command.postId);
